backend/pyproject.toml
[project]
name = "randevu-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Muhammed Ebrar",email = "kuzucuoglumuhammed@gmail.com"}
]
# readme = "README.md"  # Bu satır yoruma alındı veya silindi
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]

# BURAYI YENİDEN DÜZENLEYİN:
##packages = [{include = "randevu_backend"}] # Artık randevu_backend klasörünü paket olarak belirtiyoruz.

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"